Move the strategy logic into a new component
